[#49] ref: CategoryViewController.swift 파일 분리  - 코드는 변경하거나 추가 된 사항은 없습니다  - CategoryViewController 의 setup** , subscribe** 메서드들을 extension하여 파일을 별도로 분리했습니다.
